+++ addons/main/magazines/CfgMagRifleMisc.hpp
@@ -7,6 +7,12 @@ class rhsgref_30Rnd_792x33_SmE_StG: CA_Magazine {
 
 class rhsgref_24rnd_75x55_Stgw57: CA_Magazine {
     mass = 20;
+
+    #if __has_include("\hlc_wp_sigamt\config.bin")
+        model = "\hlc_wp_sigamt\mesh\magazine\magazine_stgw.p3d";
+        modelSpecial = "\hlc_wp_sigamt\mesh\magazine\Proxy\24rnd_GP11_STGW57";
+        modelSpecialIsProxy = 1;
+    #endif
 };
 
 class rhsgref_50Rnd_792x57_SmE_drum: CA_Magazine {